requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.2
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
et_xmlfile==2.0.0
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
openpyxl==3.1.5
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
whitenoise==6.11.0
yarl==1.22.0
